+++ CHANGELOG.md
@@ -5,15 +5,16 @@ All notable changes to PGS will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). Dates are *YYYY-MM-DD*.
 
-## [1.1.0] – 2021-xx-xx
+## [1.1.0] – 2021-06-xx
 
 ### Added
 
-
+- New Class: `PGS_CirclePacking` — circle packings of shapes, subject to varying constraints and patterns of tangencies
 - `closestPointPair()` to `PGS_Optimisation`. The method efficiently computes the **closest** pair of points in a set of points.
 - `farthestPointPair()` to `PGS_Optimisation`. The method efficiently computes the **farthest** pair of points in a set of points.
-- `chaikinCut()` to `PGS_Morphology`. The method smooths shapes via iterated corner cuts.
+- `chaikinCut()` to `PGS_Morphology`. The method smoothes shapes via iterated corner cuts.
 - `createHeart()` to `PGS_Construction`. The method generates heart-shaped PShapes.
+- A new `earCutTriangulation()` method signature that takes in a PShape argument (previously, it accepted a list of points only)
 - Expand `PGS_Conversion` to support conversion between:
   - `PATH` PShape<->JTS `LineString`
   - `POINTS` PShape<->JTS `MultiPoint`
@@ -22,11 +23,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed 
 
 - Split `PGS_Processing.concaveHull()` into `concaveHullDFS()` and `concaveHullBFS()` (the method previously used the BFS approach only).
+- Compute (rather than ignore) circle sites of radius 0 (these are effectively points) in `PGS_Voronoi.voronoiCirclesDiagram()`
+- Changed the algorithm used by `PGS_Processing.generateRandomPoints()`. The new approach is ~10x faster!
 
 ### Fixed
 - Error when `concaveHull2()` was called with alpha > 1.
 - Concave hull methods no longer mutate the input point set.
 - PShapes marked as closed and having less than 3 vertices could cause an error during conversion ([#22](https://github.com/micycle1/PGS/issues/22)).
+- `PGS_Conversion.toPVector()` now handles [primitive](https://processing.org/examples/shapeprimitives.html) PShapes
+- Constrained delaunay triangulations now respect shape holes
 
 ### Removed
 - `PGS_Processing.concaveHull()` (see *Changed* above)